+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - (Feature) Backup InProgress Agency key discovery
 - (Feature) Backup & Maintenance Conditions
 - (Bugfix) Disable member removal in case of health failure
+- (Bugfix) Reorder Topology management plan steps
 
 ## [1.2.9](https://github.com/arangodb/kube-arangodb/tree/1.2.9) (2022-03-30)
 - (Feature) Improve Kubernetes clientsets management