+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (Feature) Backup & Maintenance Conditions
 - (Bugfix) Disable member removal in case of health failure
 - (Bugfix) Reorder Topology management plan steps
+- (Feature) UpdateInProgress & UpgradeInProgress Conditions
 
 ## [1.2.9](https://github.com/arangodb/kube-arangodb/tree/1.2.9) (2022-03-30)
 - (Feature) Improve Kubernetes clientsets management