+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Bugfix) Reorder Topology management plan steps
 - (Feature) UpdateInProgress & UpgradeInProgress Conditions
 - (Bugfix) Fix Maintenance switch and HotBackup race
+- (Bugfix) Fix Maintenance Condition typo
 
 ## [1.2.9](https://github.com/arangodb/kube-arangodb/tree/1.2.9) (2022-03-30)
 - (Feature) Improve Kubernetes clientsets management